requirements.txt
pandas==3.0.1
openpyxl==3.1.5
playwright==1.58.0
gspread==6.2.1
gspread-dataframe==4.0.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0
python-dotenv==1.2.2
PyQt6>=6.6.0
pytest>=8.0
